feat: finished project

+++ src/routes/memories.ts
@@ -3,12 +3,23 @@ import { z } from 'zod'
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-   app.get('/memories', async () => {
+   app.addHook('preHandler', async (req) => {
+      await req.jwtVerify();
+   })
+
+   app.get('/memories', async (req) => {
+      console.log("entrei", req.user.sub)
+
       const memories = await prisma.memory.findMany({
+         where: {
+            userId: req.user.sub,
+         },
          orderBy: {
             createdAt: 'asc',
          }
       })
+
+      console.log("memories", memories)
 
       return memories.map(memory => {
          const verifyMemoryContent =
@@ -24,7 +35,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       })
    })
 
-   app.get('/memories/:id', async (req) => {
+   app.get('/memories/:id', async (req, res) => {
       const paramsSchema = z.object({
          id: z.string(),
       })
@@ -34,6 +45,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
       const memory = await prisma.memory.findFirstOrThrow({
          where: { id }
       })
+
+      if (!memory.isPublic && memory.userId !== req.user.sub) {
+         return res.status(401).send()
+      }
 
       return memory
    })
@@ -52,14 +67,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             content,
             coverUrl,
             isPublic,
-            userId: '0eaaefff-2402-436f-9096-22791bcb333b'
+            userId: req.user.sub
          }
       })
 
       return memory
    })
 
-   app.put('/memories/:id', async (req) => {
+   app.put('/memories/:id', async (req, res) => {
       const paramsSchema = z.object({
          id: z.string(),
       })
@@ -74,7 +89,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
       const { content, coverUrl, isPublic } = bodySchema.parse(req.body)
 
-      const memory = await prisma.memory.update({
+      let memory = await prisma.memory.findFirstOrThrow({
+         where: {
+            id: req.user.sub
+         }
+      })
+
+      if (memory.id !== req.user.sub) {
+         return res.status(401).send()
+      }
+
+      memory = await prisma.memory.update({
          where: { id },
          data: {
             content,
@@ -86,12 +111,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
       return memory
    })
 
-   app.delete('/memories/:id', async (req) => {
+   app.delete('/memories/:id', async (req, res) => {
       const paramsSchema = z.object({
          id: z.string(),
       })
 
       const { id } = paramsSchema.parse(req.params)
+
+      const memory = await prisma.memory.findFirstOrThrow({
+         where: {
+            id: req.user.sub
+         }
+      })
+
+      if (memory.id !== req.user.sub) {
+         return res.status(401).send()
+      }
 
       return await prisma.memory.delete({
          where: { id }
